pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.17.2",
    "email-validator>=2.3.0",
    "fastapi>=0.127.0",
    "itsdangerous>=2.2.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.21",
    "sendgrid>=6.12.5",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
